requirements.txt
streamlit
streamlit
streamlit-webrtc
streamlit-authenticator
pdfplumber
requests
plotly
zxing
opencv-python-headless
PyYAML
bcrypt
